requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.7
pandas>=2.2.0
openpyxl>=3.1.5
pdfkit>=1.0.0
xhtml2pdf>=0.2.15
pymysql>=1.1.1
sqlalchemy>=2.0.36
alembic>=1.13.2
# Arelle headless for XBRL parsing
arelle-release>=2.25.0
